// src/lib/enums.ts
export enum UserRole {
    ADMIN = 'admin',
    MEMBER = 'member',
  }
  
  export enum FeedbackStatus {
    PENDING = 'pending',
    ADDRESSED = 'addressed',
    UNRESOLVED = 'unresolved',
  }
  
  export enum InvitationStatus {
    PENDING = 'pending',
    ACCEPTED = 'accepted',
    EXPIRED = 'expired',
  }
  